+++ src/State/FragmentState.php
@@ -7,15 +7,13 @@ namespace Rowbot\URL\State;
 use Rowbot\URL\ParserContext;
 use Rowbot\URL\String\CodePoint;
 use Rowbot\URL\String\EncodeSet;
-use Rowbot\URL\String\PercentEncodeTrait;
+use Rowbot\URL\String\PercentEncoder;
 
 /**
  * @see https://url.spec.whatwg.org/#fragment-state
  */
 class FragmentState implements State
 {
-    use PercentEncodeTrait;
-
     public function handle(ParserContext $context, string $codePoint): int
     {
         $buffer = '';
@@ -48,7 +46,8 @@ class FragmentState implements State
         }
 
         // 1.3. UTF-8 percent-encode c using the fragment percent-encode set and append the result to url’s fragment.
-        $context->url->fragment .= $this->percentEncodeAfterEncoding('utf-8', $buffer, EncodeSet::FRAGMENT);
+        $percentEncoder = new PercentEncoder();
+        $context->url->fragment .= $percentEncoder->percentEncodeAfterEncoding('utf-8', $buffer, EncodeSet::FRAGMENT);
 
         return self::RETURN_OK;
     }