+++ src/State/OpaquePathState.php
@@ -7,17 +7,17 @@ namespace Rowbot\URL\State;
 use Rowbot\URL\ParserContext;
 use Rowbot\URL\String\CodePoint;
 use Rowbot\URL\String\EncodeSet;
-use Rowbot\URL\String\PercentEncodeTrait;
+use Rowbot\URL\String\PercentEncoder;
 
 /**
  * @see https://url.spec.whatwg.org/#cannot-be-a-base-url-path-state
  */
 class OpaquePathState implements State
 {
-    use PercentEncodeTrait;
-
     public function handle(ParserContext $context, string $codePoint): int
     {
+        $percentEncoder = null;
+
         do {
             // 1. If c is U+003F (?), then set url’s query to the empty string and state to query state.
             if ($codePoint === '?') {
@@ -59,7 +59,8 @@ class OpaquePathState implements State
             // 3.3. If c is not the EOF code point, UTF-8 percent-encode c using the C0 control percent-encode set and
             // append the result to url’s path.
             if ($codePoint !== CodePoint::EOF) {
-                $context->url->path->first()->append($this->percentEncodeAfterEncoding(
+                $percentEncoder ??= new PercentEncoder();
+                $context->url->path->first()->append($percentEncoder->percentEncodeAfterEncoding(
                     'utf-8',
                     $codePoint,
                     EncodeSet::C0_CONTROL